pyproject.toml
[tool.poetry]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Madhav Girish Nair <madhavgirish02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.2"
langchain-community = "^0.3.1"
pinecone-client = "^5.0.1"
python-dotenv = "^1.0.1"
streamlit = "^1.39.0"
pymupdf = "^1.24.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
